requirements.txt
mysql-connector-python==8.0.31
mysqlclient==2.1.1
tk==0.1.0
scapy==2.4.5
ctypes